White-Box-Vulnerability-Detection-Tool
A static Web vulnerability detection tool, which detects vulnerability
in Php code.
We modified source code of a Php parser tool
(https://github.com/nikic/PHP-Parser) in order to generate output in
Json format. The parser was to used to generate syntactic tokens of the
php source files.
We Developed a tool in Java to parse the Json output, and thus identify
tainted variables and flow from source to sink statements.
Our system was able to detect sql-injection vulnerabilities and xss
attack vulnerabilities.